requirements.txt
aiogram==2.14.3
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
Babel==2.9.1
certifi==2021.5.30
chardet==4.0.0
idna==3.2
multidict==5.1.0
pip==21.2.3
python-dateutil==2.8.2
python-telegram-bot-calendar==1.0.5
pytz==2021.1
setuptools==57.4.0
six==1.16.0
typing-extensions==3.10.0.0
yarl==1.6.3
